requirements.txt
envparse>=0.2.0,<0.3.0
tortoise-orm>=0.15.1,<0.16.0
sentry-sdk>=0.14.2,<0.15.0
asyncpg>=0.20.1,<0.21.0
uvicorn>=0.11.3,<0.12.0
fastapi>=0.49.0,<0.50.0
passlib>=1.7.2,<1.8.0
bcrypt>=3.1.7,<3.2.0
pyjwt>=1.7.1,<1.8.0
